LayoutTests/media/modern-media-controls/ios-inline-media-controls/ios-inline-media-controls-layout-expected.txt

Testing the IOSInlineMediaControls layout.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS mediaControls.children.length is 4
PASS mediaControls.children[0] is mediaControls.topLeftControlsBar
PASS mediaControls.children[1] is mediaControls.centerControlsBar
PASS mediaControls.children[2] is mediaControls.bottomControlsBar
PASS mediaControls.children[3] is mediaControls.topRightControlsBar
PASS mediaControls.topLeftControlsBar.children.length is 1
PASS mediaControls.topLeftControlsBar.children[0] instanceof ButtonsContainer is true
PASS mediaControls.topLeftControlsBar.children[0].children.length is 2
PASS mediaControls.topLeftControlsBar.children[0].children[0] is mediaControls.pipButton
PASS mediaControls.topLeftControlsBar.children[0].children[1] is mediaControls.fullscreenButton
PASS mediaControls.centerControlsBar.children.length is 1
PASS mediaControls.centerControlsBar.children[0] instanceof ButtonsContainer is true
PASS mediaControls.centerControlsBar.children[0].children.length is 3
PASS mediaControls.centerControlsBar.children[0].children[0] is mediaControls.skipBackButton
PASS mediaControls.centerControlsBar.children[0].children[1] is mediaControls.playPauseButton
PASS mediaControls.centerControlsBar.children[0].children[2] is mediaControls.skipForwardButton
PASS mediaControls.bottomControlsBar.children.length is 2
PASS mediaControls.bottomControlsBar.children[0] is mediaControls.timeControl
PASS mediaControls.bottomControlsBar.children[1] is mediaControls.rightContainer
PASS mediaControls.leftContainer.children.length is 0
PASS mediaControls.timeControl.children.length is 3
PASS mediaControls.timeControl.children[0] is mediaControls.timeControl.elapsedTimeLabel
PASS mediaControls.timeControl.children[1] is mediaControls.timeControl.scrubber
PASS mediaControls.timeControl.children[2] is mediaControls.timeControl.remainingTimeLabel
PASS mediaControls.rightContainer.children.length is 3
PASS mediaControls.rightContainer.children[0] is mediaControls.airplayButton
PASS mediaControls.rightContainer.children[1] is mediaControls.tracksButton
PASS mediaControls.topRightControlsBar.children.length is 1
PASS mediaControls.topRightControlsBar.children[0] instanceof ButtonsContainer is true
PASS mediaControls.topRightControlsBar.children[0].children.length is 1
PASS mediaControls.topRightControlsBar.children[0].children[0] is mediaControls.muteButton
PASS successfullyParsed is true

TEST COMPLETE
